Organizing FANPT utils
