+++ fanpy/fanpt/containers/base.py
@@ -2,6 +2,7 @@ r"""Base class that contains the elements required to perform a FANPT calculatio
 
 from abc import ABCMeta, abstractmethod
 from fanpy.fanpt.fanpt import update_fanci_objective, linear_comb_ham
+
 import pyci
